Refactoring to support QVs in database
Code for compressing and uncompressing QVs was placed in library files
QV.c and QV.h so that the database could also manipulate compressed QV
entries.

All the commands were edited to use the common utilities in DB.c and
DB.h that is now common to all modules.

Commands have -v (verbose) options instead of -S (silent) options to be
consistent with commands in other modules.

The commands by default produce lower case DNA sequence and .fasta’s
are 80 characters wide to conform to the default set for the other
modules.