Quiva compression bug fix
When a .quiva is smallish, there is a chance of an empty Huffman code
for a value that occurs in the first 100K substation QVs.  Fixed.

Also, 1 in 100billion times the Pacbio software gives a deletion tag
letter when the deletion QV is “no deletion”.  Not a bug but the result
of compression followed by decompressing will not be equal to the
original at the deletion tag letter (which will be N in the
decompressed output).  dexqv now gives a warning and identifies the
location of said offending deletion tag letter.