Dextractor Upgrade
A minor problem with the Pacbio .bax.h5 output is that very very rarely
we see an entry where a particular position has the deletion QV for
“not possible” but the corresponding deletion tag character is not an
’N’.  This blows the property that dexqv, undexqv are inverses of each
other.  So we modified detract so that it ensures this property is true
when it pulls out a .quiva file from .bax.h5 files.

In the process I saw some IO optimization opportunities that make
detract 3x faster !