Compressor bug (pulse # too big) fixed ...
Originally pulse numbers were expected to fit in a uint16, but this is
no longer true.  dexta, undexta, dexqv, undexqv all upgraded to use and
decode int’s.  The decoders undexta and undexqv continue to work with
old files compressed with the previous versions of the compressor (we
use a new endian code to distinguish the new format).